requirements.txt
openai>=1.0.0
requests>=2.31.0
python-dotenv>=1.0.0
pytest>=8.0.0
biopython>=1.82
gprofiler-official>=1.0.0
openpyxl>=3.1.5
matplotlib>=3.8.0
pandas>=2.1.3
seaborn>=0.13.2
pronto>=1.0.0
torch
transformers
tqdm